reading ARFF header

+++ src/main/scala/Evaluator.scala
@@ -1,11 +1,12 @@
 import org.apache.flink.api.common.functions.{FlatMapFunction, MapFunction}
 import org.apache.flink.util.Collector
 
-class Evaluator extends MapFunction[(Int, Int), Double]{
+class Evaluator extends MapFunction[(Int, Int), Int]{
 
   override def map(prediction: (Int, Int)): Int = {
     val trueClass = prediction._1
     val predictedClass = prediction._2
     if (trueClass == predictedClass) 1 else 0
   }
+
 }
